// apps/worker/src/providers/fixtures.generated.ts
export const ORIGINS = [
  {
    "code": "JFK",
    "city": "New York",
    "zones": [
      "America/New_York",
      "America/Detroit",
      "America/Toronto"
    ]
  },
  {
    "code": "ORD",
    "city": "Chicago",
    "zones": [
      "America/Chicago",
      "America/Winnipeg"
    ]
  },
  {
    "code": "DEN",
    "city": "Denver",
    "zones": [
      "America/Denver",
      "America/Edmonton",
      "America/Phoenix"
    ]
  },
  {
    "code": "LAX",
    "city": "Los Angeles",
    "zones": [
      "America/Los_Angeles",
      "America/Vancouver",
      "America/Tijuana"
    ]
  },
  {
    "code": "MIA",
    "city": "Miami",
    "zones": [
      "America/Bogota",
      "America/Lima",
      "America/Panama"
    ]
  },
  {
    "code": "GRU",
    "city": "São Paulo",
    "zones": [
      "America/Sao_Paulo",
      "America/Argentina/Buenos_Aires"
    ]
  },
  {
    "code": "LHR",
    "city": "London",
    "zones": [
      "Europe/London",
      "Europe/Dublin",
      "Europe/Lisbon"
    ]
  },
  {
    "code": "CDG",
    "city": "Paris",
    "zones": [
      "Europe/Paris",
      "Europe/Brussels",
      "Europe/Madrid",
      "Europe/Amsterdam"
    ]
  },
  {
    "code": "FRA",
    "city": "Frankfurt",
    "zones": [
      "Europe/Berlin",
      "Europe/Zurich",
      "Europe/Vienna",
      "Europe/Prague",
      "Europe/Rome"
    ]
  },
  {
    "code": "DXB",
    "city": "Dubai",
    "zones": [
      "Asia/Dubai",
      "Asia/Riyadh",
      "Asia/Qatar"
    ]
  },
  {
    "code": "DEL",
    "city": "Delhi",
    "zones": [
      "Asia/Kolkata",
      "Asia/Calcutta",
      "Asia/Karachi",
      "Asia/Kathmandu"
    ]
  },
  {
    "code": "SIN",
    "city": "Singapore",
    "zones": [
      "Asia/Singapore",
      "Asia/Kuala_Lumpur",
      "Asia/Jakarta",
      "Asia/Bangkok"
    ]
  },
  {
    "code": "HKG",
    "city": "Hong Kong",
    "zones": [
      "Asia/Hong_Kong",
      "Asia/Shanghai",
      "Asia/Taipei",
      "Asia/Manila"
    ]
  },
  {
    "code": "NRT",
    "city": "Tokyo",
    "zones": [
      "Asia/Tokyo",
      "Asia/Seoul"
    ]
  },
  {
    "code": "SYD",
    "city": "Sydney",
    "zones": [
      "Australia/Sydney",
      "Australia/Melbourne",
      "Australia/Brisbane"
    ]
  },
  {
    "code": "JNB",
    "city": "Johannesburg",
    "zones": [
      "Africa/Johannesburg",
      "Africa/Nairobi",
      "Africa/Lagos"
    ]
  }
] as const;

export const CURRENCY_SYMBOL = {
  "EUR": "€",
  "USD": "$",
  "JPY": "¥",
  "MXN": "MX$",
  "GBP": "£"
} as const;